+++ drawingnote_application/lib/features/connectionPage/connection_page.dart
@@ -4,6 +4,7 @@ import '../notePage/note_page.dart';
 import '../../services/httpmanager.dart';
 import '../../services/bluetoothmanager.dart';
 import '../../datas/pagedata.dart';
+import 'paired_bluetooth_devices_widget.dart';
 
 class ConnectionPage extends StatefulWidget {
   const ConnectionPage({super.key});
@@ -101,36 +102,9 @@ class _ConnectionPageState extends State<ConnectionPage> {
                   child: const Text("Start")),
             ],
           ),
-          Column(
-            children: [
-              const Text("페어링된 블루투스 기기 목록"),
-              Expanded(
-                child: Padding(
-                  padding: const EdgeInsets.symmetric(vertical: 10),
-                  child: Container(
-                    decoration: BoxDecoration(
-                      border: Border.all(color: Colors.black, width: 2),
-                      borderRadius: BorderRadius.circular(10),
-                    ),
-                    child: SingleChildScrollView(
-                      scrollDirection: Axis.vertical,
-                      child: Column(
-                        children: [
-                          for (var device in _bluetoothmanager.devices)
-                            TextButton(
-                                onPressed: () {
-                                  _bluetoothmanager
-                                      .connectDevice(device.address);
-                                },
-                                child: Text(device.name ?? device.address))
-                        ],
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          )
+          PairedBluetoothDevicesWidget(
+            bluetoothmanager: _bluetoothmanager,
+          ),
         ],
       ),
     );
